Rename TE_Solver to te_solver
PEP8 says: Modules should have short, all-lowercase names. Underscores
can be used in the module name if it improves readability.

https://peps.python.org/pep-0008/#package-and-module-names
